+++ R/importNOAA.R
@@ -72,6 +72,29 @@
 #' "Etc/GMT-8"` for a meteorological data frame called `met`. The two data sets
 #' could then be merged based on `date`.
 #'
+#' @section Parallel Processing:
+#'
+#'   If you are importing a lot of meteorological data, this can take a long
+#'   while. This is because each combination of year and station requires
+#'   downloading a separate data file from NOAA's online data directory, and the
+#'   time each download takes can quickly add up. [importNOAA()] and
+#'   [importNOAAlite()] can use parallel processing to speed downloading up,
+#'   powered by the capable `{mirai}` package. If users have any `{mirai}`
+#'   "daemons" set, these functions will download files in parallel. The
+#'   greatest benefits will be seen if you spawn as many daemons as you have
+#'   cores on your machine, although one fewer than the available cores is often
+#'   a good rule of thumb. Your mileage may vary, however, and naturally
+#'   spawning more daemons than station-year combinations will lead to
+#'   diminishing returns.
+#'
+#'   ```
+#'   # set workers - once per session
+#'   mirai::daemons(4)
+#'
+#'   # import lots of data - NB: no change in importNOAA()!
+#'   big_met <- importNOAA(code = "037720-99999", year = 2010:2020)
+#'   ```
+#'
 #' @param code The identifying code as a character string. The code is a
 #'   combination of the USAF and the WBAN unique identifiers. The codes are
 #'   separated by a \dQuote{-} e.g. `code = "037720-99999"`.